+++ src/gui.rs
@@ -9,6 +9,7 @@ pub mod game_list;
 pub mod icon;
 pub mod ignored_items_editor;
 pub mod modal;
+pub mod notification;
 pub mod number_input;
 pub mod other_screen;
 pub mod popup_menu;