+++ CHANGELOG.md
@@ -7,6 +7,11 @@
   * Compression levels can now be customized for zip backups.
   * In addition to restoration redirects, there are now also backup redirects and bidirectional redirects.
     The redirect editor is now on the "other" screen instead of the "restore" screen.
+  * GUI: The custom games screen now has a button to preview a specific game on demand.
+    This lets you preview a custom game even if it's not yet in the backup screen's main list.
+  * GUI: When previewing a specific game on demand,
+    if it disappears from the list because save data can no longer be found for it,
+    then a notification is shown to explain what happened.
   * GUI: The "other" screen now shows when the manifest was last checked/updated.
     There is also a button to refresh on demand.
     While the manifest is updating, a small notification is displayed at the bottom of the window.